sync previous update for cvi_fb.ko 9e6a33e29
9e6a33e29:
fix display anomaly when loading cvi_fb.ko

fb and vo use the same memory. The boot logo uses vo in u-boot.
When loading cvi_fb.ko in the kernel, we do not need to use
this line to clear the buffer. Otherwise, when insmod cvi_fb.ko,
a rectangular display abnormality will appear in the middle.

Signed-off-by: carbon <carbon@milkv.io>